Divide into components
