+++ src/components/Graph/Graph2D.tsx
@@ -33,7 +33,7 @@ export const Graph2D = ({
   };
 
   const handleNodeClick = (node: NodeObject) => {
-    selectCallback(node.id.toString(), node);
+    selectCallback(String(node.id), node);
   };
 
   return (
