+++ src/components/Graph/components/ArticleInfo/index.ts
@@ -1,0 +1,1 @@
+export * from "./ArticleInfo.tsx";
