+++ src/components/Graph/components/ArticleList/index.ts
@@ -1,0 +1,1 @@
+export * from "./ArticleList.tsx";
